Moved authn test to own test file
